Move mass spectrum type override to mass spectrum view.
