+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.msd.ui/plugin.xml
@@ -8,13 +8,4 @@
             uri="fragment.e4xmi">
       </fragment>
    </extension>
-   <extension
-         point="org.eclipse.ui.preferencePages">
-      <page
-            category="org.eclipse.chemclipse.rcp.app.ui.preferences.settingsPreferencePage"
-            class="org.eclipse.chemclipse.ux.extension.msd.ui.preferences.PreferencePage"
-            id="org.eclipse.chemclipse.ux.extension.msd.ui.preferencePage"
-            name="Mass Selective Detector (MSD)">
-      </page>
-   </extension>
 </plugin>
