--- chemclipse/plugins/org.eclipse.chemclipse.ux.extension.msd.ui/src/org/eclipse/chemclipse/ux/extension/msd/ui/preferences/PreferenceSupplier.java
@@ -1,44 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2008, 2025 Lablicate GmbH.
- * 
- * All rights reserved.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v1.0 which accompanies this distribution,
- * and is available at http://www.eclipse.org/legal/epl-v10.html
- * 
- * Contributors:
- * Philip Wenig - initial API and implementation
- *******************************************************************************/
-package org.eclipse.chemclipse.ux.extension.msd.ui.preferences;
-
-import org.eclipse.chemclipse.support.preferences.AbstractPreferenceSupplier;
-import org.eclipse.chemclipse.support.preferences.IPreferenceSupplier;
-import org.eclipse.chemclipse.ux.extension.msd.ui.Activator;
-
-public class PreferenceSupplier extends AbstractPreferenceSupplier implements IPreferenceSupplier {
-
-	public static final String P_USE_PROFILE_MASS_SPECTRUM_VIEW = "useProfileMassSpectrumView";
-	public static final boolean DEF_USE_PROFILE_MASS_SPECTRUM_VIEW = false;
-
-	public static IPreferenceSupplier INSTANCE() {
-
-		return INSTANCE(PreferenceSupplier.class);
-	}
-
-	@Override
-	public String getPreferenceNode() {
-
-		return Activator.getDefault().getBundle().getSymbolicName();
-	}
-
-	@Override
-	public void initializeDefaults() {
-
-		putDefault(PreferenceSupplier.P_USE_PROFILE_MASS_SPECTRUM_VIEW, PreferenceSupplier.DEF_USE_PROFILE_MASS_SPECTRUM_VIEW);
-	}
-
-	public static boolean useProfileMassSpectrumView() {
-
-		return INSTANCE().getBoolean(P_USE_PROFILE_MASS_SPECTRUM_VIEW);
-	}
-}
